requirements.txt
beautifulsoup4==4.12.3
playwright==1.53.0
PyYAML==6.0.2
requests==2.32.5
